executor: keep using float32

+++ executor/engine/fsm_controller.go
@@ -105,8 +105,8 @@ type (
 		stateHandlers   map[fsmState]fsmStateHandler
 		stepToState     map[types.StepType]fsmState
 		defaultPids     *map[types.StepType]*types.PidSettings
-		maxDeltaHeating float64
-		minDeltaHeating float64
+		maxDeltaHeating float32
+		minDeltaHeating float32
 	}
 )
 
@@ -180,7 +180,7 @@ func (h *heatUpStateHandler) executeState() fsmState {
 
 func (h *heatUpStateHandler) enterState() {
 	h.fanPower = newConstantPowerController(h.fsm.program.ProgramSteps[h.fsm.step].Fan.Power)
-	h.heaterPower = NewPowerController(float64(h.fsm.program.ProgramSteps[h.fsm.step].TargetTemperature), &h.fsm.program.ProgramSteps[h.fsm.step].Heater, (*h.fsm.defaultPids)[types.StepTypeHeating], h.fsm.maxDeltaHeating, h.fsm.minDeltaHeating)
+	h.heaterPower = NewPowerController(float32(h.fsm.program.ProgramSteps[h.fsm.step].TargetTemperature), &h.fsm.program.ProgramSteps[h.fsm.step].Heater, (*h.fsm.defaultPids)[types.StepTypeHeating], h.fsm.maxDeltaHeating, h.fsm.minDeltaHeating)
 	h.humidifierPower = newConstantPowerController(h.fsm.program.ProgramSteps[h.fsm.step].Humidifier.Power)
 }
 
@@ -200,7 +200,7 @@ func (h *acclimateStateHandler) executeState() fsmState {
 
 func (h *acclimateStateHandler) enterState() {
 	h.fanPower = newConstantPowerController(h.fsm.program.ProgramSteps[h.fsm.step].Fan.Power)
-	h.heaterPower = NewPowerController(float64(h.fsm.program.ProgramSteps[h.fsm.step].TargetTemperature), &h.fsm.program.ProgramSteps[h.fsm.step].Heater, (*h.fsm.defaultPids)[types.StepTypeAcclimate], 0, 0)
+	h.heaterPower = NewPowerController(float32(h.fsm.program.ProgramSteps[h.fsm.step].TargetTemperature), &h.fsm.program.ProgramSteps[h.fsm.step].Heater, (*h.fsm.defaultPids)[types.StepTypeAcclimate], 0, 0)
 	h.humidifierPower = newConstantPowerController(h.fsm.program.ProgramSteps[h.fsm.step].Humidifier.Power)
 }
 
@@ -224,7 +224,7 @@ func (h *coolDownStateHandler) executeState() fsmState {
 
 func (h *coolDownStateHandler) enterState() {
 	h.fanPower = newConstantPowerController(h.fsm.program.ProgramSteps[h.fsm.step].Fan.Power)
-	h.heaterPower = NewPowerController(float64(h.fsm.program.ProgramSteps[h.fsm.step].TargetTemperature), &h.fsm.program.ProgramSteps[h.fsm.step].Heater, (*h.fsm.defaultPids)[types.StepTypeCooling], 0, 0)
+	h.heaterPower = NewPowerController(float32(h.fsm.program.ProgramSteps[h.fsm.step].TargetTemperature), &h.fsm.program.ProgramSteps[h.fsm.step].Heater, (*h.fsm.defaultPids)[types.StepTypeCooling], 0, 0)
 	h.humidifierPower = newConstantPowerController(h.fsm.program.ProgramSteps[h.fsm.step].Humidifier.Power)
 }
 
@@ -246,7 +246,7 @@ func (h *idleStateHandler) enterState() {
 	h.fsm.shutdown()
 }
 
-func newProgramFSMController(psuController *psuController, psuStatus *fsmPSUStatus, temperatures *fsmTemperatures, defaultPids *map[types.StepType]*types.PidSettings, maxDeltaHeating float64, minDeltaHeating float64) *programFSMController {
+func newProgramFSMController(psuController *psuController, psuStatus *fsmPSUStatus, temperatures *fsmTemperatures, defaultPids *map[types.StepType]*types.PidSettings, maxDeltaHeating float32, minDeltaHeating float32) *programFSMController {
 	controller := &programFSMController{
 		psuController:       psuController,
 		currentPSUStatus:    psuStatus,
